Final Adjustments
I tried to get the text in the nav to match with the alignment in the main page, but I couldn't get it to work, so I left the spacing, i added a form section in the service page, and moved the modal to the home page. attempted to fix spacing but what I tried didn't work out very well, so I undid what I did and left it like that.